+++ packages/dialect-react-ui/CHANGELOG.md
@@ -1,11 +1,19 @@
 # CHANGELOG
 
 ## [UNRELEASED]
+
+## [0.6.11] - 2022-04-26
+
 - feat: added UI in notification for email verification
 - feat: added UI in notification for sms verification
 - Add client-side address or twitter handle validation.
 - Allow thread creation using twitter handle.
 - Only show delete thread button to dialect admins.
+- chore: reintroduce wildcard exports
+
+## [0.6.10] - 2022-04-23
+
+- Bump `@dialectlabs/react` package version
 
 ## [0.6.9] - 2022-04-22